+++ src/modules/entity/components/components/HazeComponent.ts
