+++ src/modules/entity/components/components/ImageComponent.ts
