+++ src/modules/entity/components/components/KeyLightComponent.ts
