+++ src/modules/entity/components/components/ModelComponent.ts
