+++ src/modules/entity/components/components/SkyboxComponent.ts
